Move DTO serialization to model subproject
